rename unique table name
Former-commit-id: 5dd1d38e029b7e9d57e6ac141101a177d4dd5604

+++ src/source_controller/objectcontroller/service/object_unique.go
@@ -146,9 +146,23 @@ func (cli *Service) DeleteObjectUnique(req *restful.Request, resp *restful.Respo
 	cond.Field(common.BKObjIDField).Eq(objID)
 	cond.Field(common.BKOwnerIDField).Eq(ownerID)
 
+	unique := metadata.ObjectUnique{}
+	err = db.Table(common.BKTableNameObjUnique).Find(cond.ToMapStr()).One(ctx, &unique)
+	if nil != err {
+		blog.Errorf("[DeleteObjectUnique] find error: %s, raw: %#v", err, cond.ToMapStr())
+		resp.WriteError(http.StatusInternalServerError, &metadata.RespError{Msg: defErr.Error(common.CCErrObjectDBOpErrno)})
+		return
+	}
+
+	if unique.Ispre {
+		blog.Errorf("[DeleteObjectUnique] could not delete preset constrain: %s", err, unique)
+		resp.WriteError(http.StatusInternalServerError, &metadata.RespError{Msg: defErr.Error(common.CCErrObjectDBOpErrno)})
+		return
+	}
+
 	err = db.Table(common.BKTableNameObjUnique).Delete(ctx, cond.ToMapStr())
 	if nil != err {
-		blog.Errorf("[DeleteObjectUnique] Delete error: %s, raw: %#v", err)
+		blog.Errorf("[DeleteObjectUnique] Delete error: %s, raw: %#v", err, cond.ToMapStr())
 		resp.WriteError(http.StatusInternalServerError, &metadata.RespError{Msg: defErr.Error(common.CCErrObjectDBOpErrno)})
 		return
 	}
